keeping one wrapper file(called indexing) for index and indices operator under Q/OPERATORS/INDEX/ directory.
